+++ src/infrastructure/persistence/typeorm/entities/CountryEntity.ts
@@ -2,7 +2,7 @@ import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
 
 @Entity({ name: 'countries' })
 export class CountryEntity {
-    @PrimaryGeneratedColumn({ type: 'int', unsigned: true })
+    @PrimaryGeneratedColumn({ type: 'smallint', unsigned: true })
     id!: number;
 
     @Column({ type: 'varchar', length: 100, unique: true })
